refactor: GatheringReceipt 도메인 모델에서 익셉션 제거
