+++ src/main/kotlin/com/server/dpmcore/member/member/application/MemberQueryService.kt
@@ -2,7 +2,7 @@ package com.server.dpmcore.member.member.application
 
 import com.server.dpmcore.authority.domain.model.AuthorityId
 import com.server.dpmcore.member.member.application.exception.MemberNotFoundException
-import com.server.dpmcore.member.member.domain.exception.MemberTeamNotFoundException
+import com.server.dpmcore.member.member.application.exception.MemberTeamNotFoundException
 import com.server.dpmcore.member.member.domain.model.MemberId
 import com.server.dpmcore.member.member.domain.model.query.MemberNameAuthorityQueryModel
 import com.server.dpmcore.member.member.domain.port.inbound.MemberQueryByAuthorityUseCase
